BERTopic/requirements.txt
pandas==2.2.2
bertopic==0.16.0
sentence-transformers==2.6.1
scikit-learn==1.4.2
umap-learn==0.5.5
matplotlib==3.8.4
numpy==1.26.4
hdbscan